start producer from main.py and handle ws there
