+++ telemetry-producer/producer/producer.py
@@ -1,17 +1,9 @@
-import websocket
-import time
+
 import numpy as np
-
-try:
-    import thread
-except ImportError:
-    import _thread as thread
-import time
-
 from lsst.ts import salobj
-import time
+
 import json
-import os
+
 from utils import NumpyEncoder
 import importlib
 
@@ -143,44 +135,3 @@ class Producer:
             "category": "event",
             "data": output_dict
         }))
-
-def on_ws_open(ws,send_ws_data):
-    def run(*args):
-        while True:
-            time.sleep(2)
-            send_ws_data(ws)
-            
-        time.sleep(1)
-        ws.close()
-        print("thread terminating...")
-    thread.start_new_thread(run, ())
-    print("open")
-
-def on_ws_message(ws, message):
-    print("### message ###")
-    print(message)
-
-def on_ws_error(ws, error):
-    print("### error ###")
-    print(error)
-
-def on_ws_close(ws):
-    print("### closed ###")
-
-if __name__=='__main__':
-
-    producer = Producer()
-
-    WS_HOST = os.environ["WEBSOCKET_HOST"]
-    websocket.enableTrace(True)
-    ws = websocket.WebSocketApp("ws://%s/" % WS_HOST,
-                            on_message = on_ws_message,
-                            on_error = on_ws_error,
-                            on_close = on_ws_close)
-
-    ws.on_open = lambda ws: on_ws_open(ws, producer.send_ws_data)
-
-    #Emitter
-    while True:
-        time.sleep(3)
-        ws.run_forever()
